+++ PruebaNETCore/Models/PeriodosMantenimiento.cs
@@ -7,7 +7,7 @@ public partial class PeriodosMantenimiento
 {
     public int Id { get; set; }
 
-    public decimal PeriodoKilometraje { get; set; }
+    public int PeriodoKilometraje { get; set; }
 
     public virtual ICollection<Vehiculo> Vehiculos { get; set; } = new List<Vehiculo>();
 }
